requirements.txt
playwright
beautifulsoup4
pandas
python-dotenv
requests
schedule
rich
